% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC_climatefile.R
\name{PWC_fg}
\alias{PWC_fg}
\title{Weather file for Pesticide in Water Calculator (PWC) v1.52 and Pesticide Root Zone Model (PRZM5) v5.02}
\usage{
PWC_fg(data, date, start, end, temp_celsius, precip_cm, ws10_cm_s, pevp_cm,
  solr_lang, save_in)
}
\arguments{
\item{data}{Name of the dataset}

\item{date}{Column name for dates in format}

\item{start}{Date to start the weather file in format}

\item{end}{Date to end the weather file in format}

\item{temp_celsius}{Column name for temperature (Celsius)}

\item{precip_cm}{Column name for precipitation (cm/day)}

\item{ws10_cm_s}{Column name for wind speed values (cm/sec)}

\item{pevp_cm}{Column name for pan pan evaporation data (cm/day)}

\item{solr_lang}{Column name for solar radiation (Langley)}

\item{save_in}{Path to save the final weather file. Extension .dvf do not need to be specified.}
}
\value{
A weather file for PWC and PRZM5 models
}
\description{
The function \code{\link{PWC_fg}} generates the weather files for \href{https://www.epa.gov/pesticide-science-and-assessing-pesticide-risks/models-pesticide-risk-assessment}{PWC and PRZM5} models from a database in .csv format. The weather file format is described in the \href{}{PRZM5 manual} guidelines.
}
\examples{
plot_crayons()

}
