% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC_fg.R
\name{PWC_fg}
\alias{PWC_fg}
\title{Weather file for Pesticide in Water Calculator (PWC) and Pesticide Root Zone Model (PRZM5)}
\usage{
PWC_fg(data, date, start, end, format_date, temp_celsius, precip_cm,
  ws10_cm_s, pevp_cm, solr_lang, save_in)
}
\arguments{
\item{data}{Name of the dataset.}

\item{date}{Column name for dates}

\item{start}{Date to start the weather file}

\item{end}{Date to end the weather file}

\item{format_date}{Date format (more details in \href{https://www.rdocumentation.org/packages/base/versions/3.6.1/topics/strptime}{strptime} documentation)}

\item{temp_celsius}{Column name for temperature (Celsius)}

\item{precip_cm}{Column name for precipitation (cm/day)}

\item{ws10_cm_s}{Column name for wind speed values (cm/sec)}

\item{pevp_cm}{Column name for panevaporation data (cm/day)}

\item{solr_lang}{Column name for solar radiation (Langley)}

\item{save_in}{Path to save the final weather file. Extension .dvf do not need to be specified.}
}
\value{
To create a weather file in a format readable by PWC and PRZM5 models
}
\description{
The function \code{\link{PWC_fg}} converts a weather dataset to the file needed as input for PWC and PRZM5.
Units should be converted previously. The format is described in detail in the \href{http://bit.ly/2k6yV26}{PRZM5 manual}.
}
\examples{
# Example of a small dataset.
data <- data.frame(date = c("01/01/81", "02/01/81", "03/01/81", "04/01/81"),
                   precip = c(0.00, 0.10, 0.00, 0.00),
                   evap = c(0.30, 0.21, 0.28, 0.28),
                   tmed = c(9.5, 6.3, 3.5, 5),
                   wind = c(501.6, 368.0, 488.3, 404.5),
                   solrad = c(240.3, 244.3, 303.0, 288.5))

# Load the package
library(PWCfilegenerator, warn.conflicts = FALSE)

# Function to convert and save the weather file as a .dvf
PWCfilegenerator::PWC_fg(data = data,
                        date = "date",
                        format = "\%d/\%m/\%y",
                        start ="02/01/81",
                        end = "03/01/81",
                        precip_cm = "precip",
                        pevp_cm = "evap",
                        temp_celsius = "tmed",
                        ws10_cm_s = "wind",
                        solr_lang = "solrad",
                        save_in = "F:/folder/filename") # The extension .dvf will be added automatically.


}
